+++ addon/background.js
@@ -60,13 +60,13 @@ chrome.action.onClicked.addListener(() => {
   });
 });
 chrome.commands?.onCommand.addListener((command) => {
-  if (command !== "open-popup"){
+  if (!command.startsWith("open-")){
     chrome.tabs.create({
       url: `chrome-extension://${chrome.i18n.getMessage("@@extension_id")}/${command}.html?host=${sfHost}`
     });
   } else {
     chrome.runtime.sendMessage({
-      msg: "shortcut_pressed", sfHost
+      msg: "shortcut_pressed", command, sfHost
     });
   }
 });